callback改成accesstoken

+++ passport-main-service/src/main/java/com/sogou/upd/passport/service/connect/AccessTokenService.java
@@ -19,7 +19,7 @@ public interface AccessTokenService {
      * @return
      * @throws ServiceException
      */
-    public boolean initialOrUpdateAccessToken(String userid,String Accesstoken,int expire) throws ServiceException;
+    public boolean initialOrUpdateAccessToken(String userid,String Accesstoken,long expire) throws ServiceException;
 
     /**
      * 通过userid获取accesstoken
